refactor(diff): migrate to Component-based rendering, split large files
- Create DiffComponent: proper pi-tui Component with render(width),
  Box wrapper for self-contained background, lazy Shiki highlighting

- Split ansi.ts (328loc) into ansi/ directory:
  codes.ts (constants, interfaces, legacy aliases),
  manip.ts (string utilities), colors.ts (theme derivation)

- Split render.ts (419loc) into render/ directory:
  shared.ts (helpers, config), unified.ts, split.ts
  Functions now accept width param and return string[]

- Split index.ts tools into tools/ directory:
  write.ts (write tool with DiffComponent renderResult),
  edit.ts (edit tool with Box shell, renderShell: self)

- renderSplit/renderUnified get backward-compat wrappers plus
  width-aware renderSplitLines/renderUnifiedLines variants

- DiffBg colors derived per-instance instead of global mutable state
