+++ packages/diff/src/tools/index.ts
@@ -1,0 +1,4 @@
+/** Tool registration — barrel re-export. */
+
+export { registerWriteTool } from "./write.js";
+export { registerEditTool } from "./edit.js";
